% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmentation.R
\name{augmentation}
\alias{augmentation}
\title{augmentation 

Description augmentation}
\usage{
augmentation(n = seq(0, 10, by = 0.01), s = 1, d = 5, aMin = 1,
  aMax = 2)
}
\arguments{
\item{n}{n}

\item{s}{s}

\item{d}{d}

\item{aMin}{aMin}

\item{aMax}{aMax}
}
\value{
augmentation
}
