% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carpenterMod.R
\name{carpenterMod}
\alias{carpenterMod}
\title{Carpenter lake model

Runs simulation of the Carpenter Lake model}
\usage{
carpenterMod(x0, tf = 100, a = 1, b = 1, r = 1, FUN, verbose = FALSE)
}
\arguments{
\item{x0}{x0}

\item{tf}{tf}

\item{a}{a}

\item{b}{b}

\item{r}{r}

\item{FUN}{FUN}

\item{verbose}{verbose}
}
\value{
carpenterMod
}
