% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagplot.R
\name{lagplot}
\alias{lagplot}
\title{lagplot 

Description lagplot}
\usage{
lagplot(x, k = 1, xlab, ylab, type = "l", std = FALSE, add = FALSE,
  col = "grey", pch = 19, cex = 1, lwd = 1)
}
\arguments{
\item{x}{x}

\item{k}{k}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{type}{type}

\item{std}{std}

\item{add}{add}

\item{col}{col}

\item{pch}{pch}

\item{cex}{cex}

\item{lwd}{lwd}
}
\value{
lagplot
}
