% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mineralization.R
\name{mineralization}
\alias{mineralization}
\title{mineralization 

Description mineralization}
\usage{
mineralization(x, w, c = 100)
}
\arguments{
\item{x}{x}

\item{w}{w}

\item{c}{c}
}
\value{
mineralization
}
