% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseSimulator.R
\name{noiseSimulator}
\alias{noiseSimulator}
\title{noiseSimulator 

Description noiseSimulator}
\usage{
noiseSimulator(x0, tf = 100, a = 1, b = 1, r = 1, eta = 0, FUN, NOISE,
  verbose = FALSE)
}
\arguments{
\item{x0}{x0}

\item{tf}{tf}

\item{a}{a}

\item{b}{b}

\item{r}{r}

\item{eta}{eta}

\item{FUN}{FUN}

\item{NOISE}{NOISE}

\item{verbose}{verbose}
}
\value{
noiseSimulator
}
