% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/photo.R
\name{photo}
\alias{photo}
\title{photo 

Description photo}
\usage{
photo(days = 3, Amax = 1, Amin = 0, Aqe = 0.3, LCP = 0, start = 0,
  amp = 50)
}
\arguments{
\item{days}{days}

\item{Amax}{Amax}

\item{Amin}{Amin}

\item{Aqe}{Aqe}

\item{LCP}{LCP}

\item{start}{start}

\item{amp}{amp}
}
\value{
photo
}
