% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppHyst.R
\name{ppHyst}
\alias{ppHyst}
\title{ppHyst 

Description ppHyst}
\usage{
ppHyst(x, n1, n2, feedingTime = 720, tol = 0)
}
\arguments{
\item{x}{x}

\item{n1}{n1}

\item{n2}{n2}

\item{feedingTime}{feedingTime}

\item{tol}{tol}
}
\value{
ppHyst
}
