% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppReturn.R
\name{ppReturn}
\alias{ppReturn}
\title{ppReturn 

Description ppReturn}
\usage{
ppReturn(x, feed.time = 720, thresh = 1e-05, minutes = FALSE)
}
\arguments{
\item{x}{x}

\item{thresh}{thresh}

\item{minutes}{minutes}
}
\value{
ppReturn
}
