% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppSim.R
\name{ppSim}
\alias{ppSim}
\title{ppSim 

Description ppSim}
\usage{
ppSim(days = 5, foodWeight = 1, beta = 4.5e-06, d = 5, k = 2,
  Amax = 20, Amin = Amax/1.5, m = 0, aMax = 2, aMin = 1, s = 1,
  feedingTime = 720, c = 1, x0 = 0, w0 = 0, w.w = 75, photo.val = 1,
  Ascalar = 0.75, Bscalar = 1, bound.max = FALSE, verbose = FALSE,
  decycle = FALSE)
}
\arguments{
\item{days}{days}

\item{foodWeight}{foodWeight}

\item{beta}{beta}

\item{d}{d}

\item{k}{k}

\item{Amax}{Amax}

\item{Amin}{Amin}

\item{m}{m}

\item{aMax}{aMax}

\item{aMin}{aMin}

\item{s}{s}

\item{feedingTime}{feedingTime}

\item{c}{c}

\item{x0}{x0}

\item{w0}{w0}

\item{Ascalar}{Ascalar}

\item{Bscalar}{Bscalar}

\item{verbose}{verbose}

\item{decycle}{decycle}

\item{w}{w}

\item{photo}{photo}

\item{bound}{bound}
}
\value{
ppSim
}
