\name{SarModel-package}
\alias{SarModel-package}
\alias{SarModel}
\docType{package}
\title{
\packageTitle{SarModel}
}
\description{
\packageDescription{SarModel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SarModel}
\packageIndices{SarModel}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{SarModel}

Maintainer: \packageMaintainer{SarModel}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
