% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPrey.R
\name{simPrey}
\alias{simPrey}
\title{simPrey

Description simPrey}
\usage{
simPrey(days = 30, prey.mass = 10, prey.rate = 3, perturb.mass = 10,
  perturb.rate = 3)
}
\arguments{
\item{days}{days}

\item{prey}{prey}

\item{prey}{prey}

\item{perturb}{perturb}

\item{perturb}{perturb}
}
\value{
simPrey
}
